+++ src/ti/backups.c
@@ -571,7 +571,7 @@ int ti_backups_backup(void)
 
     do
     {
-        ti_sleep(100);
+        (void) ti_sleep(100);
 
         uv_mutex_lock(backups->lock);